fixes imports from pages package rename updates logger config
