+++ ab_testing_tool_app/urls.py
@@ -1,14 +1,14 @@
 from django.conf.urls import patterns, include, url
 from django.contrib import admin
 
-from ab_testing_tool_app.pages.main_pages import (render_stage_control_panel, not_authorized,
+from ab_testing_tool_app.views.main_pages import (render_stage_control_panel, not_authorized,
      tool_config)
 
-from ab_testing_tool_app.pages.selection_pages import (resource_selection, submit_selection,
+from ab_testing_tool_app.views.selection_pages import (resource_selection, submit_selection,
                     submit_selection_new_stage)
-from ab_testing_tool_app.pages.stage_pages import (create_stage, submit_create_stage,
+from ab_testing_tool_app.views.stage_pages import (create_stage, submit_create_stage,
                     edit_stage,  delete_stage, deploy_stage, submit_edit_stage)
-from ab_testing_tool_app.pages.track_pages import (create_track, submit_create_track,
+from ab_testing_tool_app.views.track_pages import (create_track, submit_create_track,
                     submit_edit_track, edit_track, delete_track)
 
 admin.autodiscover()
