+++ ab_testing_tool_app/views/selection_pages.py
@@ -37,7 +37,6 @@ def submit_selection(request):
     stage_id = request.REQUEST.get("stage_id")
     t = Stage.objects.get(pk=stage_id)
     page_url = stage_url(request, stage_id)
-    print stage_id, page_url
     page_name = t.name
     content_return_url = request.REQUEST.get("content_return_url")
     params = {"return_type": "lti_launch_url",
